refactor: extract types in its own file
