+++ rules/sort-imports/types.ts
@@ -1,0 +1,62 @@
+import type {
+  DeprecatedCustomGroupsOption,
+  PartitionByCommentOption,
+  SpecialCharactersOption,
+  NewlinesBetweenOption,
+  FallbackSortOption,
+  GroupsOptions,
+  OrderOption,
+  RegexOption,
+  TypeOption,
+} from '../../types/common-options'
+import type { SortingNode } from '../../types/sorting-node'
+
+export type Options = [
+  Partial<{
+    customGroups: {
+      value?: DeprecatedCustomGroupsOption
+      type?: DeprecatedCustomGroupsOption
+    }
+    partitionByComment: PartitionByCommentOption
+    specialCharacters: SpecialCharactersOption
+    locales: NonNullable<Intl.LocalesArgument>
+    newlinesBetween: NewlinesBetweenOption
+    fallbackSort: FallbackSortOption
+    internalPattern: RegexOption[]
+    groups: GroupsOptions<Group>
+    environment: 'node' | 'bun'
+    partitionByNewLine: boolean
+    sortSideEffects: boolean
+    tsconfigRootDir?: string
+    maxLineLength?: number
+    ignoreCase: boolean
+    order: OrderOption
+    type: TypeOption
+    alphabet: string
+  }>,
+]
+
+export type Group =
+  | 'side-effect-style'
+  | 'external-type'
+  | 'internal-type'
+  | 'builtin-type'
+  | 'sibling-type'
+  | 'parent-type'
+  | 'side-effect'
+  | 'index-type'
+  | 'internal'
+  | 'external'
+  | 'sibling'
+  | 'unknown'
+  | 'builtin'
+  | 'parent'
+  | 'object'
+  | 'index'
+  | 'style'
+  | 'type'
+  | string
+
+export interface SortImportsSortingNode extends SortingNode {
+  isIgnored: boolean
+}
